Выносит функции отрисовки из utils в utils/render
